Split Gulp tasks across multiple files
Close #874
