+++ gulp/build-webpack.js
@@ -1,0 +1,4 @@
+import gulp from 'gulp';
+import webpackBuild from '../webpack/build';
+
+gulp.task('build-webpack', ['env'], webpackBuild);
