+++ gulp/build.js
@@ -1,0 +1,3 @@
+import gulp from 'gulp';
+
+gulp.task('build', ['build-webpack']);
