+++ gulp/clean.js
@@ -1,0 +1,4 @@
+import del from 'del';
+import gulp from 'gulp';
+
+gulp.task('clean', () => del('build/*'));
