+++ gulp/eslint.js
@@ -1,0 +1,4 @@
+import gulp from 'gulp';
+import runEslint from './support/run-eslint.js';
+
+gulp.task('eslint', () => runEslint());
