+++ gulp/mocha.js
@@ -1,0 +1,6 @@
+import gulp from 'gulp';
+import mochaRunCreator from '../test/mochaRunCreator';
+
+gulp.task('mocha', () => {
+  mochaRunCreator('process')();
+});
