+++ gulp/server-hot.js
@@ -1,0 +1,4 @@
+import bg from 'gulp-bg';
+import gulp from 'gulp';
+
+gulp.task('server-hot', bg('node', './webpack/server'));
