+++ gulp/server-node.js
@@ -1,0 +1,4 @@
+import bg from 'gulp-bg';
+import gulp from 'gulp';
+
+gulp.task('server-node', bg('node', './src/server'));
