+++ gulp/support/args.js
@@ -1,0 +1,7 @@
+import yargs from 'yargs';
+
+const args = yargs
+  .alias('p', 'production')
+  .argv;
+
+export default args;
