+++ gulp/support/run-eslint.js
@@ -1,0 +1,20 @@
+import args from './args';
+import eslint from 'gulp-eslint';
+import gulp from 'gulp';
+import gulpIf from 'gulp-if';
+
+// To fix some eslint issues: gulp eslint --fix
+const runEslint = () => {
+  const isFixed = file => args.fix && file.eslint && file.eslint.fixed;
+  return gulp.src([
+    'gulpfile.babel.js',
+    'messages/*.js',
+    'src/**/*.js',
+    'webpack/*.js'
+  ], { base: './' })
+    .pipe(eslint({ fix: args.fix }))
+    .pipe(eslint.format())
+    .pipe(gulpIf(isFixed, gulp.dest('./')));
+};
+
+export default runEslint;
