+++ gulp/test.js
@@ -1,0 +1,6 @@
+import gulp from 'gulp';
+import runSequence from 'run-sequence';
+
+gulp.task('test', done => {
+  runSequence('eslint-ci', 'mocha', 'build-webpack', done);
+});
